+++ CountlyServerConfig.m
@@ -38,6 +38,7 @@
 @property (nonatomic) NSInteger version;
 @property (nonatomic) long long timestamp;
 @property (nonatomic) long long lastFetchTimestamp;
+@property (nonatomic) BOOL serverConfigDisabled;
 
 @end
 
@@ -108,12 +109,16 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
         _version = 0;
         _currentServerConfigUpdateInterval = 4;
         _requestTimer = nil;
+        _serverConfigDisabled = NO;
     }
     return self;
 }
 
 - (void)retrieveServerConfigFromStorage:(NSString *)sdkBehaviorSettings
 {
+    if (_serverConfigDisabled) {
+        return;
+    }
     NSError *error = nil;
     NSDictionary *serverConfigObject = [CountlyPersistency.sharedInstance retrieveServerConfig];
     if (serverConfigObject.count == 0 && sdkBehaviorSettings)
@@ -197,6 +202,10 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
 
 - (void)notifySdkConfigChange:(CountlyConfig *)config
 {
+    if (_serverConfigDisabled) {
+        return;
+    }
+    
     config.enableDebug = _loggingEnabled || config.enableDebug;
     CountlyCommon.sharedInstance.enableDebug = config.enableDebug;
 
@@ -297,6 +306,10 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
 
 - (void)fetchServerConfigIfTimeIsUp
 {
+    if (_serverConfigDisabled) {
+        return;
+    }
+    
     if (_lastFetchTimestamp)
     {
         long long currentTime = NSDate.date.timeIntervalSince1970 * 1000;
@@ -311,7 +324,11 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
 
 - (void)fetchServerConfig:(CountlyConfig *)config
 {
+    if (_serverConfigDisabled) {
+        return;
+    }
     CLY_LOG_D(@"Fetching server configs...");
+    
     if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
         return;
 
@@ -397,6 +414,10 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
     }
 
     CLY_LOG_D(@"serverConfigRequest URL :%@", URL);
+}
+
+- (void)disableSDKBehaviourSettings {
+    _serverConfigDisabled = YES;
 }
 
 - (BOOL)trackingEnabled