+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## XX.XX.XX
 * The feedback widgets now have fullscreen and transparent backgrounds for a cleaner look.
+* Added a config method to disable server config in the initialization "disableSDKBehaviorSettings()".
 
 ## 25.4.1
 * Mitigated an issue that could occur while serializing events to improve stability, performance and memory usage.